Trieudo (#25)
* add holiday controller

* Develop (#14)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

* merge main (#12)

* synchronize response APIs (#10)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

* adjust db and clean code (#11)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

---------


* #merge main to develop (#13)

* synchronize response APIs (#10)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

* adjust db and clean code (#11)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

---------


* change url

* company && employee features

* adjust employee apis - update only reference field

* adjust get all contract api with paging response

---------


* Develop (#15)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

* merge main (#12)

* synchronize response APIs (#10)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

* adjust db and clean code (#11)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

---------


* #merge main to develop (#13)

* synchronize response APIs (#10)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

* adjust db and clean code (#11)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

---------


* change url

* company && employee features

* adjust employee apis - update only reference field

* adjust get all contract api with paging response

---------


* add holiday crud

* fix naming conventions and response paging api (#20)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

* merge main (#12)

* synchronize response APIs (#10)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

* adjust db and clean code (#11)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

---------


* #merge main to develop (#13)

* synchronize response APIs (#10)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

* adjust db and clean code (#11)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

---------


* change url

* company && employee features

* adjust employee apis - update only reference field

* adjust get all contract api with paging response

* Merge main into Devlop (#18)

* Develop (#14)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

* merge main (#12)

* synchronize response APIs (#10)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

* adjust db and clean code (#11)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

---------


* #merge main to develop (#13)

* synchronize response APIs (#10)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

* adjust db and clean code (#11)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

---------


* change url

* company && employee features

* adjust employee apis - update only reference field

* adjust get all contract api with paging response

---------


* Develop (#15)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

* merge main (#12)

* synchronize response APIs (#10)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

* adjust db and clean code (#11)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

---------


* #merge main to develop (#13)

* synchronize response APIs (#10)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

* adjust db and clean code (#11)

* add api get profile and fix preload in contract api

* update contract api

* synchronize the response apis

* adjust db and clean code

* adjust db and clean code

---------


* change url

* company && employee features

* adjust employee apis - update only reference field

* adjust get all contract api with paging response

---------


---------


* fix naming conventions

---------


* update holiday controller

* add salaryType controller

* add authorize

* update controller

* add payroll controller

* add payroll controller

---------

+++ SalaryManagement.Api/Controllers/HolidayController.cs
@@ -12,6 +12,7 @@ namespace SalaryManagement.Api.Controllers
 {
     [Route("api/v1/holidays")]
     [ApiController]
+    [Authorize]
     public class HolidayController : ControllerBase
     {
         private readonly IHolidayService _holidayService;
@@ -23,10 +24,11 @@ namespace SalaryManagement.Api.Controllers
             _mapper = mapper;
         }
 
-        [HttpGet("getAllHolidays")]
+        [HttpGet("get-all")]
         public async Task<IActionResult> GetAllHoliday()
         {
             var holidays = await _holidayService.GetAllHoliday();
+
             return Ok(holidays);
         }
 
@@ -34,36 +36,84 @@ namespace SalaryManagement.Api.Controllers
         public async Task<IActionResult> FindById(string id)
         {
             var holiday = await _holidayService.GetHolidaysById(id);
+            if(holiday == null)
+            {
+                return NotFound("Holiday not found");
+            }
+            else
+            {
             return Ok(holiday);
         }
+        }
 
-        [HttpPost("add")]
-        public async Task<IActionResult> AddHoliday(Holiday holiday)
+        [HttpPost("")]
+        public async Task<IActionResult> AddHoliday(HolidayRequest request)
         {
+            await Task.CompletedTask;
+            string id = Guid.NewGuid().ToString();
+
+            Holiday holiday = new Holiday
+            {
+                HolidayId = id,
+                StartDate = request.StartDate,
+                EndDate = request.EndDate,
+                IsDeleted = true
+            };
             var result = await _holidayService.AddHoliday(holiday);
+
             return Ok(result);
         }
 
-        [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteHoliday(string id)
+        [HttpDelete("delete")]
+        public async Task<IActionResult> DeleteHoliday(HolidayDelete request)
         {
+            string id = request.Id;
+            var msg = "";
+
+            Holiday holiday = await _holidayService.GetHolidaysById(id);
+
+            if (holiday != null)
+            {
             var result = await _holidayService.DeleteHoliday(id);
-            if(result is null)
+                if (result)
+                {
+                    msg = "Delete successfully";
+                }
+                else
+                {
+                    msg = "Delete failed";
+                }
+            } 
+            else
             {
-                return NotFound("Holiday not found!!!");
+                msg = "Holiday not found";
             }
-            return Ok(result);
+            return Ok(msg);
         }
 
-        [HttpPut("{id}")]
-        public async Task<IActionResult> UpdateHoliday(string id, Holiday request)
+        [HttpPut("update")]
+        public async Task<IActionResult> UpdateHoliday(HolidayUpdate request)
         {
-            var result = await _holidayService.UpdateHoliday(id, request);
-            if(result is null)
+            Holiday holiday = new Holiday
             {
-                return NotFound("Holiday not found!!!");
+                HolidayId = request.Id,
+                StartDate = request.StartDate,
+                EndDate = request.EndDate,
+                IsDeleted = request.IsDelete
+            };
+
+            var result = await _holidayService.UpdateHoliday(holiday);
+            var msg = "";
+
+            if (result)
+        {
+                msg = "Update successfully";
             }
-            return Ok(result);
+            else
+            {
+                return NotFound("Holiday not found");
+            }
+            return Ok(msg);
         }
     }
 }
